Remove support for CPUID on ia64
Forwarded: by email to <gary@issiweb.com>, CC to <lunar@debian.org> on 2016-06-27

The current implementation simply fail to build and the 'cpuid' instruction on
Itanium spreads out value differently than what 'cpuid' does on i386 and amd64.
So the code needs to be worked out quite differently to work.

In the meantime, disabling CPUID support will make haveged use the default
cache values on ia64.

Gbp-Pq: Name 0001-Remove-support-for-CPUID-on-ia64.patch