Implement a command mode and use it for chroot
the current only known command is root=<new_root> which enforce
an already running haveged to switch by using chroot(2) to new
root file system and to an execv(3) there to become rebased.

This allows haveged to survive the switch root within initrd
and hold the entropy at high level.

Signed-off-by: Werner Fink <werner@suse.de>